+++ src/components/DracCard.ts
@@ -10,13 +10,14 @@ const styles = styled(
     --cardBackground:#EFEFEF;
     --disabledColor:#D8D8D8;
     --textColor:black;
+    --puffinCardBackground:var(--explorerItemSelectedBackground);
     transition:0.1s;
     font-family:var(--puffinFont,var(--font));
     padding:10px 20px;
     min-height:30px;
     min-width:30px;
     /*max-width:300px;*/
-    max-height:300px;
+    /*max-height:300px;*/
     background:var(--puffinCardBackground,var(--cardBackground));
     border:0;
     outline:0;
@@ -28,18 +29,18 @@ const styles = styled(
     display:inline-block;
     overflow:auto;
 }
-:host:hover:not(:active){
+:host.clickable:hover:not(:active){
     cursor:pointer;
     transition:0.1s;
     box-shadow:0px 1px 5px 2px rgba(0,0,0,.05) ,0 0 0 0px var(--puffinAccent,var(--accentColor));
 }
-:host:active{
+:host.clickable:active{
     border:0;
     outline:0;
     box-sizing: border-box;
     box-shadow:0px 1px 5px 3px rgba(0,0,0,.05) ,0 0 0 3px var(--puffinAccent,var(--accentColor));
 }
-:host.active{
+:host.clickable.active{
     border:0;
     outline:0;
     box-sizing: border-box;
@@ -47,7 +48,7 @@ const styles = styled(
 }
 :host.disabled{
     pointer-events: none;
-    background:var(--puffinDisabledColor,var(--disabledColor));
+    /*background:var(--puffinDisabledColor,var(--disabledColor));*/
 }
 `
 )
@@ -56,13 +57,16 @@ export const DracCard = ({
   width = "auto",
   children,
   onclick,
+  clickable = false,
 }: {
   width: string | number
   children: any[]
   onclick?: Function
+  clickable: boolean
 }) => {
   return html`
-      <${styles} style="${`width:${width};`}" onclick=${() => (onclick ? onclick() : null)}>
+      <${styles} style="${`width:${width};`}" class="${onclick ? "clickable" : ""}" onclick=${() =>
+    onclick ? onclick() : null}>
         ${children}
       </${styles}>
     `